requirements.txt
Django==3.1
djangorestframework==3.12.1
psycopg2==2.8.4
ipython==7.8.0
xhtml2pdf==0.2.4